refactor: persistence 경로 변경
